requirements.txt
streamlit==1.25.0
PyPDF2==3.0.1
PyCryptodome==3.18.0
python-magic==0.4.27
spacy==3.6.0
neo4j==5.11.0
pyvis==0.3.2
pandas==2.0.3
numpy==1.24.3
python-dotenv==1.0.0
boto3==1.28.57
botocore==1.31.57
awscli-local==0.20
httpx==0.24.1
openai==1.3.5
anthropic==0.7.4